# Threading the VCF annotator