@bwalsh 2025-04-01
@david4096 Hey! Good to hear from you.  

We did some integration a while ago and noticed the same thing - threading helps.   We added threading to our wrapper.   I'm curious what parameters (# of threads etc) you used and how much it helped?


https://docs.google.com/presentation/d/1YUTGW3CaXimUE44aMEe9DpP1qN_mESoL7guqDXTweYo/edit#slide=id.p
https://docs.google.com/presentation/d/1hk-c2T2w6X2sh5Dlwqzxi9kjR-c91R2_snyh1q0XCLI/edit#slide=id.p