@david4096 2025-04-01
Hi @bwalsh !! I put in a cursory benchmark in the above issue. The code here uses your CPU count, which for me was 12. It took a little less than half the time (which I'm sure could be improved upon).